resolve step bindings for the current scope
The binding function for step definitions (Given, When, Then), will now
look up the step binding from the registry using the step pattern and tags
that are currently in scope. If a step binding cannot be found, or if
multiple step bindings are found, then an appropriate error is thrown.

The BindingRegistry will now only look for the default tag if no step
bindings could be found with the tag list provided.